requirements.txt
anyio==3.6.1
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
boto3==1.23.10
botocore==1.26.10
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==2.2.28
django-allauth==0.50.0
django-bootstrap4==22.1
django-cron==0.6.0
django-crontab==0.7.1
django-environ==0.8.1
django-extensions==3.1.5
django-filter==2.4.0
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.13.1
drf-yasg==1.20.3
factory-boy==3.2.1
Faker==13.15.1
google-api-core==2.8.2
google-auth==2.6.6
google-cloud-core==2.3.2
google-cloud-secret-manager==2.10.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.48.0
grpcio-status==1.48.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
oauthlib==3.2.0
onesignal-sdk==2.0.0
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
portalocker==2.5.1
proto-plus==1.20.6
protobuf==4.21.4
psycopg2==2.8.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.1
pywin32==304
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
sendgrid==6.8.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
stripe==3.5.0
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
waitress==2.1.2
whitenoise==6.0.0
